Decoupling hive-frontend from neo4j
