+++ packages/backends/hive-frontend/src/server.ts
@@ -1,13 +1,103 @@
 import { HiveFrontendServer } from ".";
 import { createServer } from 'http'
 import express from 'express';
-
+import { driver, auth } from 'neo4j-driver'
 const greenlock = require("greenlock-express");
 
 (async (port: number = 8000) => {
 	const app = express()
 
-	const frontendServer = new HiveFrontendServer({routes: []})
+
+	const neoDriver = driver(
+		  process.env.NEO4J_URI || "neo4j://localhost",
+		  auth.basic(
+			process.env.NEO4J_USER || "neo4j",
+			process.env.NEO4J_PASSWORD || "test"
+		  )
+		);
+	const neoSession = neoDriver.session();
+	
+
+	  
+	const frontendServer = new HiveFrontendServer({
+		routes: [],
+		getUser: async (profile: { id: string }) => {
+
+			const session = neoDriver?.session();
+			const data = await session?.run(`
+			  MATCH (org:HiveOrganisation)-[:TRUSTS]->(user:HiveUser {id: $id})
+			  CALL {
+				WITH user
+				MATCH (user)-[:HAS_ROLE]->()-->(apps:HiveAppliance)
+				RETURN distinct(apps{.*}) as apps
+			  }
+			  RETURN user{
+				id: user.id,
+				name: user.name,
+				organisation: org.id,
+				applications: collect(apps{.*})
+			  }
+			`, {
+			  
+				id: profile.id,
+			  
+			})
+			  
+			const user = data.records?.[0].get(0);
+			console.log("deserializeUser", user);
+			session.close()
+			return user; 			
+		},
+		getViews: async (req) => {
+
+			const session = neoDriver?.session()
+
+			const apps = await session?.readTransaction(async (tx) => {
+			  let apps: any[] = [];
+		
+			  if (!req || !req.user.id) {
+				const result = await tx?.run(`
+						MATCH (apps:HiveAppliance)
+						WHERE apps.entrypoint IS NOT NULL
+						RETURN distinct(apps{.*})
+					`);
+		
+				apps = result?.records.map((x) => x.get(0)) || [];
+			  } else {
+				const result = await tx?.run(
+				  `
+						MATCH (user:HiveUser {id: $id})-[:HAS_ROLE]->()-->(apps:HiveAppliance)
+						WHERE apps.entrypoint IS NOT NULL
+						RETURN distinct(apps{.*})
+					`,
+				  {
+					id: req.user.id,
+				  }
+				);
+		
+				apps = result?.records.map((x) => x.get(0)) || [];
+			  }
+			  return apps || [];
+			}) || [];
+			
+			session?.close()
+
+			const views = (apps || []).map((app) => ({
+				name: app.name,
+				path: app.slug,
+				default: false,
+			}))
+
+			const appliances = apps?.map((app) => ({
+			  name: app.name,
+			  config_url: app.entrypoint,
+			}))
+
+			return { views, apps: appliances }
+
+		
+		}
+	})
 
 	app.use(frontendServer.connect)
 	const server = createServer(app)
